% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_gof.R
\name{adf_gof.class}
\alias{adf_gof.class}
\title{An S4 class to represent the Goodness-of-Fit of the Angular Dependence Function estimates}
\usage{
adf_gof.class(adf, ray, blocksize, nboot, alpha, gof)
}
\description{
An S4 class to represent the Goodness-of-Fit of the Angular Dependence Function estimates
}
\section{Slots}{

\describe{
\item{\code{adf}}{An S4 object of class \code{adf_est.class}.}

\item{\code{ray}}{\loadmathjax{} Ray \mjeqn{\omega}{} to be considered on the goodness of fit assessment.}

\item{\code{blocksize}}{Size of the blocks for the block bootstrap procedure. If \code{1} (default), then a standard bootstrap approach is applied.}

\item{\code{nboot}}{Number of bootstrap samples to be taken. Default is \code{250} samples.}

\item{\code{alpha}}{Significance level to compute the \mjeqn{(1-\alpha)}{}\% confidence intervals. Default is \code{0.05}.}

\item{\code{gof}}{A list containing the model and empirical exponential quantiles, and the lower and upper bound of the confidence interval.}
}}

\keyword{internal}
