% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_gof.R
\docType{methods}
\name{adf_gof}
\alias{adf_gof}
\title{Diagnostic tool for the ADF estimates \eqn{\lambda(\omega)}}
\usage{
adf_gof(
  data,
  w_ind,
  w = seq(0, 1, by = 0.01),
  lambda,
  q = 0.95,
  blocksize = 1,
  nboot = 250,
  alpha = 0.05
)
}
\arguments{
\item{data}{matrix that contains the data, in standard exponential margins}

\item{w_ind}{index of the ray to be considered while assessing the estimates}

\item{w}{sequence of angles between 0 and 1; default set to a vector of 101 equally spaced angles}

\item{lambda}{estimates of the angular dependence function}

\item{q}{quantile to be used for the Hill estimator and/or the Heffernan and Tawn conditional extremes model; default set to 0.95}

\item{blocksize}{size of the blocks for the block bootstrap; default to 1 for a standard bootstrap approach}

\item{nboot}{number of bootstrap samples; default to 250}

\item{alpha}{significance level to compute the confidence intervals}
}
\value{
return a list containing the model and empirical quantiles, and the lower and upper bounds for the confidence interval
}
\description{
computes the goodness of fit of the angular dependence function estimates \eqn{\lambda(\omega)}
}
\details{
to do
}
\examples{
library(ReturnCurves)

}
\references{
to do
}
