% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_est.R
\docType{methods}
\name{adf_est}
\alias{adf_est}
\title{Estimation of the Angular Dependence function (ADF)}
\usage{
adf_est(
  data,
  w = seq(0, 1, by = 0.01),
  method = c("hill", "cl"),
  qhill = 0.95,
  qalphas = 0.95,
  k = 7,
  constrained = FALSE
)
}
\arguments{
\item{data}{A matrix or data frame containing the data in standard exponential margins.}

\item{w}{Sequence of angles between 0 and 1. Default is \code{seq(0, 1, by = 0.01)}.}

\item{method}{String that indicates which method is used for the estimation of the angular dependence function. Must either be \code{"hill"}, to use the Hill estimator \insertCite{Hill1975}{ReturnCurves}, or \code{"cl"} to use the composite likelihood estimator approaches. More details can be found in \code{\link{adf_est}}.}

\item{qhill}{Marginal quantile used for the Hill estimator \insertCite{Hill1975}{ReturnCurves}. Default is 0.95.}

\item{qalphas}{Marginal quantile used for the Heffernan and Tawn conditional extremes model \insertCite{HeffernanTawn2004}{ReturnCurves}. Default set to 0.95.}

\item{k}{Polynomial degree for the Bernstein-Bezier polynomials used for the estimation of the angular dependence function using the composite likelihood method. Default set to 7.}

\item{constrained}{Logical. If FALSE (default) no knowledge of the conditional extremes parameters is incorporated in the angular dependence function estimation.}
}
\value{
A vector containing the estimates of the angular dependence function.
}
\description{
\loadmathjax{}
Estimation of the angular dependence function \mjeqn{\lambda(\omega)}{} introduced by \insertCite{WadsworthTawn2013;textual}{ReturnCurves}.
}
\details{
\loadmathjax{} The angular dependence function \mjeqn{\lambda(\omega)}{} can be estimated through a pointwise estimator, obtained with the Hill estimator, or via a smoother approach, 
obtained using Composite likelihood methods. Knowledge of the conditional extremes framework introduced by \insertCite{HeffernanTawn2004;textual}{ReturnCurves} can be incorporated by setting \code{"constrained"} to \code{TRUE}.
For more details see \insertCite{MurphyBarltropetal2024;textual}{ReturnCurves}.
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(100), runif(100))

dataexp <- margtransf(data)

w <- seq(0, 1, by = 0.01)

adf <- adf_est(data = dataexp, method = "hill")

\dontrun{
plot(w, pmax(w, 1-w), type = "l", lty = 2)
lines(w, adf, col = 2, lwd = 2)
}

}
\references{
\insertAllCited{}
}
