% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airdata.R
\docType{data}
\name{airdata}
\alias{airdata}
\title{Air pollution data}
\format{
Data frame with 1427 observations (rows) and 2 variables (columns):
\describe{
\item{nox}{Vector containing daily measurements of NOx, in parts per billion.}
\item{pm10}{Vector containing daily measurements of PM10, in ug/m3.}
}
}
\source{
\code{airdata} is a subset of the data provided in the \code{\link[openair]{openair}} R package.
}
\description{
Air pollution data from Marylebone, London (UK), between December and February 
of 1998 to 2005. It contains daily measurements of air pollutant concentrations of NOx and
PM10. The dataset is a subset of the data from the \code{openair} package and
consists of \code{1427} observations.
}
\keyword{datasets}
