% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{wads_tawn_curve_exp}
\alias{wads_tawn_curve_exp}
\title{Return curve}
\usage{
wads_tawn_curve_exp(data_exp, prob, q, w = seq(0.001, 0.999, length.out = 200))
}
\arguments{
\item{data_exp}{to do}

\item{prob}{to do}

\item{q}{to do}

\item{w}{to do}
}
\value{
to do
}
\description{
to do
}
\examples{
library(ReturnCurves)

}
