% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marggpd.R
\name{marggpd.class}
\alias{marggpd.class}
\title{An S4 class to represent the assessment of the the Marginal Tail Fits}
\usage{
marggpd.class(margdata, blocksize, nboot, alpha, marggpd)
}
\description{
An S4 class to represent the assessment of the the Marginal Tail Fits
}
\section{Slots}{

\describe{
\item{\code{margdata}}{An S4 object of class \code{margtransf.class}. See \code{\link{margtransf}} for more details.}

\item{\code{blocksize}}{Size of the blocks for the block bootstrap procedure. If \code{1} (default), then a standard bootstrap approach is applied.}

\item{\code{nboot}}{Number of bootstrap samples to be taken. Default is \code{250} samples.}

\item{\code{alpha}}{Significance level to compute the \mjeqn{(1-\alpha)}{}\% confidence intervals. Default is \code{0.05}.}

\item{\code{gof}}{A list containing the model and empirical exponential quantiles, and the lower and upper bound of the confidence interval.}
}}

\keyword{internal}
