% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpd_diag.R
\docType{methods}
\name{marggpd}
\alias{marggpd}
\title{Assessing the Generalised Pareto Distribution Fit}
\usage{
marggpd(margdata, nboot = 250, blocksize = 1, alpha = 0.05)
}
\arguments{
\item{margdata}{An S4 object of class \code{margtransf.class}. See \code{\link{margtransf}} for more details.}

\item{nboot}{Number of bootstrap samples to be taken. Default is \code{250} samples.}

\item{blocksize}{Size of the blocks for the block bootstrap procedure. If \code{1} (default), then a standard bootstrap approach is applied.}

\item{alpha}{Significance level to compute the \mjeqn{(1-\alpha)}{}\% tolerance intervals. Default is \code{0.05}.}
}
\value{
An object of S4 class \code{marggpd.class}. This object returns the arguments of the function and an extra slot \code{marggpd} which is a list containing: 
\item{model}{A list containing the model quantiles for each variable.} 
\item{empirical}{A list containing the empirical quantiles for each variable.}
\item{lower}{A list containing the lower bounds of the confidence interval for each variable.}
\item{upper}{A list containing the upper bounds of the confidence interval for each variable.}
}
\description{
\loadmathjax{}
Assessment of the generalised Pareto distribution fit for each margin after following the marginal transformation procedure \code{\link{margtransf}}.
}
\details{
Let \mjeqn{F^{-1}_{GPD}}{} denote the inverse of the cumulative distribution function of a variable following a Generalised Pareto Distribution (GPD) and \mjeqn{X_{(i)}}{}, \mjeqn{Y_{(i)}}{}denote the \mjeqn{i}{i}-th ordered increasing statistics, \mjeqn{i = 1, \ldots, n}{}. 
Function \code{plot} shows QQ plots between the model and empirical exponential quantiles for both variables, i.e. points \mjeqn{\left(F^{-1}_{GPD}\left(\frac{i}{n+1}\right), X_{(i)}\right)}{} and \mjeqn{\left(F^{-1}_{GPD}\left(\frac{i}{n+1}\right), Y_{(i)}\right)}{},
along with the line \mjeqn{y=x}{}. Uncertainty is obtained via a (block) bootstrap procedure and shown by the grey region on the plot.
A good fit is shown by agreement of model and empirical quantiles, i.e. points should lie close to the line \mjeqn{y=x}{}. 
In addition, line \mjeqn{y = x}{} should mainly lie within the \mjeqn{(1-\alpha)}{}\% tolerance intervals.
}
\examples{
library(ReturnCurves)

data(airdata)

n <- dim(airdata)[1]

margdata <- margtransf(airdata)

marggpd <- marggpd(margdata = margdata)

plot(marggpd)

\dontrun{
# To see the the S4 object's slots
str(marggpd)

# To access the list of vectors
marggpd@marggpd
}

}
\references{
\insertAllCited{}
}
