% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_transform.R
\name{margtransf.class}
\alias{margtransf.class}
\title{An S4 class to represent the Marginal Transformation}
\usage{
margtransf.class(data, qmarg, dataexp)
}
\description{
An S4 class to represent the Marginal Transformation
}
\section{Slots}{

\describe{
\item{\code{data}}{A matrix containing the data on the original margins.}

\item{\code{qmarg}}{A vector containing the marginal quantile used to fit the Generalised Pareto Distribution (GPD) for each variable. Default is \code{rep(0.95, 2)}.}

\item{\code{dataexp}}{A matrix containing the data on standard exponential margins.}
}}

\keyword{internal}
