% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_est.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot.adf}
\title{Visualisation of the ADF estimates}
\usage{
\S4method{plot}{adf_est.class}(x)
}
\arguments{
\item{x}{An object of an adf_est S4 class produced by \code{\link{adf_est}}.}
}
\value{
A ggplot object.
}
\description{
Plot method for an S4 object returned by \code{\link{adf_est}}.
}
\details{
\loadmathjax{} The plot shows a comparison between the estimates \mjeqn{\hat{\lambda}(\omega)}{} of the ADF and its theoretical lower bound.
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(1000), rnorm(1000))

dataexp <- margtransf(data)

w <- seq(0, 1, by = 0.01)

lambda <- adf_est(data = dataexp, method = "hill")

plot(lambda)

}
