% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_est.R
\docType{methods}
\name{plot,adf_est.class,ANY-method}
\alias{plot,adf_est.class,ANY-method}
\alias{plot,adf_est.class}
\title{Visualisation of the Angular Dependence Function estimates}
\usage{
\S4method{plot}{adf_est.class,ANY}(x)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{adf_est}}.}
}
\value{
A ggplot object showing a comparison between the Angular Dependence Function (ADF) estimates and the lower bound max\mjeqn{\lbrace \omega, 1-\omega\rbrace}{}.
}
\description{
Plot method for an S4 object returned by \code{\link{adf_est}}.
}
\keyword{internal}
