% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_gof.R
\docType{methods}
\name{plot,adf_gof.class,ANY-method}
\alias{plot,adf_gof.class,ANY-method}
\alias{plot,adf_gof.class}
\title{Visualisation of the Goodness-of-Fit of the Angular Dependence Function estimates}
\usage{
\S4method{plot}{adf_gof.class,ANY}(x)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{adf_gof}}.}
}
\value{
A ggplot object showing the QQ-plot between the model and empirical exponential quantiles.
}
\description{
Plot method for an S4 object returned by \code{\link{adf_gof}}.
}
\keyword{internal}
