% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_est.R
\docType{methods}
\name{plot,rc_est.class,ANY-method}
\alias{plot,rc_est.class,ANY-method}
\alias{plot,rc_est.class}
\title{Visualisation of the Return Curve estimates}
\usage{
\S4method{plot}{rc_est.class,ANY}(x)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{rc_est}}.}
}
\value{
A ggplot object showing the original data with the estimated Return Curve.
}
\description{
Plot method for an S4 object returned by \code{\link{rc_est}}.
}
\keyword{internal}
