% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_gof.R
\docType{methods}
\name{plot,rc_gof.class,ANY-method}
\alias{plot,rc_gof.class,ANY-method}
\alias{plot,rc_gof.class}
\title{Visualisation of the Goodness-of-Fit of the Return Curve estimates}
\usage{
\S4method{plot}{rc_gof.class,ANY}(x)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{rc_gof}}.}
}
\value{
A ggplot object comparing the true and median estimates of the empirical probability of lying in a survival region.
}
\description{
Plot method for an S4 object returned by \code{\link{rc_gof}}.
}
\keyword{internal}
