% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_uncertainty.R
\docType{methods}
\name{rc_unc}
\alias{rc_unc}
\title{Uncertainty of the Return Curve estimates}
\usage{
rc_unc(
  data,
  w = seq(0, 1, by = 0.01),
  p,
  method = c("hill", "cl"),
  qmarg = 0.95,
  q = 0.95,
  qalphas = 0.95,
  k = 7,
  constrained = FALSE,
  blocksize = 1,
  nboot = 250,
  nangles = 150,
  alpha = 0.05
)
}
\arguments{
\item{data}{A matrix or a data frame which contains the data in the original margins.}

\item{w}{Sequence of angles between 0 and 1. Default is \code{seq(0, 1, by = 0.01)}.}

\item{p}{Curve survival probability.}

\item{method}{String that indicates which method is used for the estimation of the angular dependence function. Must either be "hill", to use the Hill estimator, or "cl" to use the composite likelihood estimator approaches. More details can be found in \code{\link{adf_est}}.}

\item{qmarg}{Marginal quantile used to fit the Generalised Pareto Distribution. Default is 0.95.}

\item{q}{Marginal quantile used for the min-projection variable and Hill estimator. Default is 0.95}

\item{qalphas}{Marginal quantile used for the Heffernan and Tawn conditional extremes model (see ). Default set to 0.95}

\item{k}{Polynomial degree for the Bernstein-Bezier polynomials used for the estimation of the angular dependence function using the composite likelihood method. Default set to 7.}

\item{constrained}{Logical. If FALSE (default) no knowledge of the conditional extremes parameters is incorporated in the angular dependence function estimation.}

\item{blocksize}{Size of the blocks for the block bootstrap procedure. If 1, then a standard bootstrap approach is applied.}

\item{nboot}{Number of bootstrap samples to be taken. Default is 250 samples.}

\item{nangles}{Number of angles \eqn{m} in the \eqn{(0, \pi/2)} interval (see ). Default is 150 angles.}

\item{alpha}{Significance level to compute the \eqn{(1-\alpha)} confidence intervals. Default is 0.05.}
}
\value{
Returns a list containing: \describe{
\item{median}{A vector containing the median estimates of the return curve.} 
\item{mean}{A vector containing the mean estimates of the return curve.} 
\item{lower}{A vector containing the lower bound of the confidence interval.}
\item{upper}{A vector containing the upper bound of the confidence interval.}
}
}
\description{
Uncertainty assessment of the return curve estimates following .
}
\details{
to do
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(100), runif(100))

dataexp <- margtransf(data)

prob <- 1e^-3

rc <- rc_est(data = dataexp, p = prob, method = "hill")

rc_orig <- curvetransf(curvedata = rc, data = data)

unc <- rc_unc(data = data, p = prob, method = "hill")


\dontrun{
plot(data, xlab = "X", ylab = "Y", pch = 16)
lines(rc_orig, lwd = 2, col = 2)
lines(unc$median, lwd = 2, col = "orange") # to plot median estimates
lines(unc$mean, lwd = 2, col = "orange") # to plot mean estimates
lines(unc$upper, lty = 'dashed', col = 'blue', lwd = 2)
lines(unc$lower, lty = 'dashed', col = 'blue', lwd = 2)
}

}
\references{
\insertRef{MurphyBarltropetal2023}{ReturnCurves}, \insertRef{HeffernanTawn2024}{ReturnCurves}
}
