% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvetransf.R
\docType{methods}
\name{curvetransf}
\alias{curvetransf}
\title{Return Curve estimates on original margins}
\usage{
curvetransf(curvedata, data, qmarg = 0.95)
}
\arguments{
\item{curvedata}{A matrix containing an object of function \code{\link{rc_est}}.}

\item{data}{A matrix containing the data on the original margins.}

\item{qmarg}{Marginal quantile to be used for the fit of the Generalised Pareto Distribution. Default is 0.95.}
}
\value{
A matrix containing the estimates of the Return Curve on the original margins.
}
\description{
Back transformation of the return curve estimates onto the original margins following \insertCite{MurphyBarltropetal2023;textual}{ReturnCurves}.
}
\note{
The parameter \code{qmarg} should be the same as the one used in \code{\link{margtransf}}.
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(100), runif(100))

dataexp <- margtransf(data)

prob <- 0.001

rc <- rc_est(data = dataexp, p = prob, method = "hill")

rc_orig <- curvetransf(curvedata = rc, data = data)

\dontrun{
plot(data, pch = 20, main = "Return Curve in original margins")
lines(rc_origin, col = 2, lwd = 2)
}

}
\references{
\insertAllCited{}
}
