% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnCurves-package.R
\docType{package}
\name{ReturnCurves-package}
\alias{ReturnCurves}
\alias{ReturnCurves-package}
\title{ReturnCurves: Estimation of the Return Curves}
\description{
\loadmathjax{}
Estimation of the angular dependence function \mjeqn{\lambda(\omega)}{} \insertCite{WadsworthTawn2013}{ReturnCurves}, and of the Return Curves \insertCite{MurphyBarltropetal2023}{ReturnCurves}
}
\section{Available functions}{
 
\code{\link{adf_est}}: Estimation of the Angular Dependence Function (ADF)

\code{\link{adf_gof}}: Goodness of fit of the Angular Dependence Function estimates

\code{\link{margtransf}}: Marginal Transformation

\code{\link{rc_est}}: Return Curve estimation

\code{\link{rc_gof}}: Goodness of fit of the Return Curve estimates

\code{\link{rc_unc}}: Uncertainty of the Return Curve estimates
}

\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(1000), rnorm(1000))

n <- dim(data)[1]

# Marginal Transformation
margdata <- margtransf(data)

head(margdata@dataexp)

# Return Curves estimation

prob <- 1/n

retcurve <- rc_est(margdata = margdata, p = prob, method = "hill")

head(retcurve@rc)

# ADF estimation
lambda <- adf_est(margdata = margdata, method = "hill")

head(lambda@adf)

}
\references{
\insertAllCited{}
}
\author{
\strong{Maintainer}: Lídia André \email{l.andre@lancaster.ac.uk}

Authors:
\itemize{
  \item Callum Murphy-Barltrop \email{callum.murphy-barltrop@tu-dresden.de}
}

}
