% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\docType{methods}
\name{runShiny}
\alias{runShiny}
\title{Complementary Shiny app for the ReturnCurves package}
\usage{
runShiny()
}
\description{
Launches the R Shiny app complementary to the \code{\link{ReturnCurves-package}}.
}
