% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_gof.R
\docType{methods}
\name{adf_gof}
\alias{adf_gof}
\title{Goodness of fit of the Angular Dependence function estimates}
\usage{
adf_gof(adf, w_ind, blocksize = 1, nboot = 250, alpha = 0.05)
}
\arguments{
\item{adf}{An S4 object of class \code{adf_est.class}. See \code{\link{adf_est}} for more details.}

\item{w_ind}{Index of the ray to be considered on the goodness of fit assessment.}

\item{blocksize}{Size of the blocks for the block bootstrap procedure. If \code{1} (default), then a standard bootstrap approach is applied.}

\item{nboot}{Number of bootstrap samples to be taken. Default is \code{250} samples.}

\item{alpha}{\loadmathjax{}Significance level to compute the \mjeqn{(1-\alpha)}{}\% confidence intervals. Default is \code{0.05}.}
}
\value{
An object of S4 class \code{adf_gof.class}. This object returns the arguments of the function and an extra slot \code{gof} which is a list containing: 
\item{model}{A vector containing the model quantiles.} 
\item{empirical}{A vector containing the empirical quantiles.}
\item{lower}{A vector containing the lower bound of the confidence interval.}
\item{upper}{A vector containing the upper bound of the confidence interval.}
}
\description{
\loadmathjax{}
Assessment of the goodness of fit of the angular dependence function estimates \mjeqn{\lambda(\omega)}{} following the procedure of \insertCite{MurphyBarltropetal2024;textual}{ReturnCurves}.
}
\details{
\loadmathjax{} Define the min-projection variable as \mjeqn{t^1_\omega = t_\omega - u_\omega | t_\omega > u_\omega}{}, then
variable \mjeqn{\lambda(\omega)T^1_\omega \sim Exp(1)}{} as \mjeqn{u_\omega \to \infty}{} for all \mjeqn{\omega \in [0,1]}{}. 
A good fit is shown by agreement of model and empirical quantiles, i.e. points should lie close to the line \mjeqn{y=x}{} (lie within the \mjeqn{(1-\alpha)}{} confidence band).
}
\note{
\loadmathjax{} It is recommended to assess the goodness-of-fit of \mjeqn{\lambda(\omega)}{} for a few values of \mjeqn{\omega}{w}.
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(1000), rnorm(1000))

dataexp <- margtransf(data)@dataexp

lambda <- adf_est(data = dataexp, method = "hill")

w_ind <- 31
gof <- adf_gof(adf = lambda, w_ind = w_ind)

plot(gof)

\dontrun{
# To see the the S4 object's slots
str(gof)

# To access the list of vectors
gof@gof
}

}
\references{
\insertAllCited{}
}
