% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_gof.R
\docType{methods}
\name{adf_gof}
\alias{adf_gof}
\title{Goodness of fit of the ADF estimates \eqn{\lambda(\omega)}}
\usage{
adf_gof(
  data,
  w_ind,
  w = seq(0, 1, by = 0.01),
  lambda,
  q = 0.95,
  blocksize = 1,
  nboot = 250,
  alpha = 0.05
)
}
\arguments{
\item{data}{A matrix or a data frame that contains the data in standard exponential margins.}

\item{w_ind}{Index of the ray to be considered on the goodness of fit assessment.}

\item{w}{Sequence of angles between 0 and 1. Default is \code{seq(0, 1, by = 0.01)}.}

\item{lambda}{Vector containing the estimates of the angular dependence function \eqn{\lambda{\omega}}.}

\item{q}{Marginal quantile to be used for the min-projection variable \eqn{T^*} (see Details). Default is 0.95.}

\item{blocksize}{Size of the blocks for the block bootstrap procedure. If 1, then a standard bootstrap approach is applied.}

\item{nboot}{Number of bootstrap samples to be taken. Default is 250 samples.}

\item{alpha}{Significance level to compute the \eqn{(1-\alpha)} confidence intervals. Default is 0.05.}
}
\value{
Returns a list containing: \describe{
\item{model}{A vector containing the model quantiles.} 
\item{empirical}{A vector containing the empirical quantiles.}
\item{lower}{A vector containing the lower bound of the confidence interval.}
\item{upper}{A vector containing the upper bound of the confidence interval.}
}
}
\description{
Assessment of the goodness of fit of the angular dependence function estimates \eqn{\lambda(\omega)} following the procedure of \insertCite{MurphyBarltropetal2024}{ReturnCurves}.
}
\details{
Define the min projection variable as \eqn{t*_\omega = t_\omega - u_\omega | t_\omega > u_\omega}.  
Variable \eqn{\lambda(\omega)T^*_\omega \sim Exp(1)} as \eqn{u_\omega \to \infty} for all \eqn{\omega \in [0,1].} 
Therefore, a good agreement between the model and empirical quantiles should occur and a the \eqn{y=x} between the quantiles should lie within the \eqn{(1-\alpha)} confidence band.
The lower and upper bounds of the confidence interval are obtained through bootstrapp of the min-projection variable.
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(100), runif(100))

dataexp <- margtransf(data)

lambda <- adf_est(data = dataexp, method = "hill")

w_ind <- 31
gof <- adf_gof(data = dataexp, w_ind = w_ind, lambda = lambda)

\dontrun{
plot(gof$model, gof$empirical, ylab = "Empirical", xlab = "Model")
polygon(c(rev(gof$model), gof$model), c(rev(gof$lower), gof$upper), col = 'grey', border = NA)
points(gof$model, gof$empirical, pch = 16, col = "black")
abline(0, 1, col = 2,lwd = 3) 
}

}
\references{
\insertRef{MurphyBarltropetal2024}{ReturnCurves}
}
