% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lamb.R
\docType{methods}
\name{est_lamb}
\alias{est_lamb}
\title{Estimation of the Angular Dependence function \eqn{\lambda(\omega)}}
\usage{
est_lamb(
  data,
  w,
  method = c("hill", "cl"),
  q = 0.95,
  k = 7,
  constrained = "no"
)
}
\arguments{
\item{data}{matrix that contains the data, in standard exponential margins}

\item{w}{sequence of angles between 0 and 1; default set to a vector of 101 equally spaced angles}

\item{method}{method to be used in the estimation of the angular dependence function: "hill" to use the Hill estimator, "cl" for the composite likelihood estimator}

\item{q}{quantile to be used for the Hill estimator and/or the Heffernan and Tawn conditional extremes model; default set to 0.95}

\item{k}{polynomial degree for the Bernstein-Bezier polynomials used in the estimation of the angular dependence function using the composite likelihood method; default set to 7}

\item{constrained}{indicates whether or not to incorporate knowledge of the conditional extremes parameters; default set to "no"}
}
\value{
adf estimates
}
\description{
estimates the angular dependence function \eqn{\lambda(\omega)} ... to do
}
\details{
This function returns the estimation of the Angular Dependence Function (ADF) denoted by \eqn{\lambda(\omega)} given in ...

This can be estimated through a pointwise estimator, obtained with the Hill estimator, or via a smooth approach, 
obtained using Composite likelihood methods. Knowledge of the conditional extremes framework introduced by Heffernan 
and Tawn can be incorporated by setting the argument "contrained" to "no"
}
\examples{
library(ReturnCurves)

}
\references{
to do
}
