% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_transform.R
\docType{methods}
\name{margtransf}
\alias{margtransf}
\title{Marginal Transformation}
\usage{
margtransf(data, q = 0.95)
}
\arguments{
\item{data}{matrix that contains the data, in the original margins}

\item{q}{quantile to be used for the GPD; default set to 0.95}
}
\value{
matrix containing the variables in standard exponential margins
}
\description{
computes the marginal transformation to standard exponential through the probability integral transform... to do
}
\examples{
library(ReturnCurves)

}
