% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_transform.R
\docType{methods}
\name{margtransf}
\alias{margtransf}
\title{Marginal Transformation}
\usage{
margtransf(data, q = 0.95)
}
\arguments{
\item{data}{A matrix or data frame containing the data in original margins.}

\item{q}{Marginal quantile used to fit the Generalised Pareto Distribution. Default is 0.95.}
}
\value{
A matrix or data frame containing the data in standard exponential margins.
}
\description{
Marginal transformation of the random vector to standard exponential margins, following \insertCite{m2023}.
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(100), runif(100))

dataexp <- margtransf(data)

}
\references{
\insertRef{MurphyBarltropetal2023}{m2023}
}
