% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_transform.R
\docType{methods}
\name{margtransf}
\alias{margtransf}
\title{Marginal Transformation}
\usage{
margtransf(data, qmarg = 0.95)
}
\arguments{
\item{data}{A matrix containing the data on the original margins.}

\item{qmarg}{Marginal quantile used to fit the Generalised Pareto Distribution (GPD). Default is \code{0.95}.}
}
\value{
An object of S4 class \code{margtransf.class}. This object returns the arguments of the function and extra slot \code{dataexp} containing a matrix with the data on standard margins. To visualise the joint distribution, set \code{joint = T} in the \code{plot} environment. If \code{joint = F} (default), \code{plot} returns the marginal distribution of each variable.
}
\description{
Marginal transformation of a random vector to standard exponential margins following \insertCite{ColesTawn1991;textual}{ReturnCurves}.
}
\details{
\loadmathjax{} Given a threshold value \mjeqn{u}{u}, a stationary random vector 
is transformed by using the empirical cumulative distribution function 
(cdf) below \mjeqn{u}{u}, and a GPD fit above \mjeqn{u}{u}.
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(1000), rnorm(1000))

dataexp <- margtransf(data)

# Plots the marginal distributions of X and Y on original vs standard exponential margins
plot(dataexp) 

# Plots the joint distribution of X and Y on original vs standard exponential margins
plot(dataexp, joint = T) 

}
\references{
\insertAllCited{}
}
