% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marggpd.R
\docType{methods}
\name{plot,marggpd.class,ANY-method}
\alias{plot,marggpd.class,ANY-method}
\alias{plot,marggpd.class}
\title{Visualisation of the assessment of the Marginal Tail Fits}
\usage{
\S4method{plot}{marggpd.class,ANY}(x)
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{marggpd}}.}
}
\value{
A ggplot object showing the QQ-plot between the model and empirical generalised Pareto distribution quantiles.
}
\description{
Plot method for an S4 object returned by \code{\link{marggpd}}.
}
\keyword{internal}
