% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_transform.R
\docType{methods}
\name{plot,margtransf.class,ANY-method}
\alias{plot,margtransf.class,ANY-method}
\alias{plot,margtransf.class}
\title{Visualisation of the Marginal Transformation}
\usage{
\S4method{plot}{margtransf.class,ANY}(x, which = c("all", "hist", "ts", "joint"))
}
\arguments{
\item{x}{An instance of an S4 class produced by \code{\link{margtransf}}.}

\item{which}{String that indicates which type of plot to show. Must either be \code{"all"} (Default), \code{"hist"}, \code{"ts"} or \code{"joint"}.}
}
\value{
A ggplot object showing:
\item{\code{which = "hist"}}{histograms of each variable on original and standard exponential margins.}
\item{\code{which = "ts"}}{time series of each variable on original and standard exponential margins.}
\item{\code{which = "joint"}}{joint distribution on original and standard exponential margins.}
\item{\code{which = "all"}}{all the available plots.}
}
\description{
Plot method for an S4 object returned by \code{\link{margtransf}}.
}
\keyword{internal}
