% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_gof.R
\docType{methods}
\name{rc_gof}
\alias{rc_gof}
\title{Diagnostic tool for the Return Curve Estimates}
\usage{
rc_gof(
  data,
  w = seq(0, 1, by = 0.01),
  method = c("hill", "cl"),
  rc_origin,
  q = 0.95,
  k = 7,
  constrained = "no",
  blocksize = 1,
  nboot = 250,
  nangles = 150,
  alpha = 0.05
)
}
\arguments{
\item{data}{matrix that contains the data, in standard exponential margins}

\item{w}{sequence of angles between 0 and 1; default set to a vector of 101 equally spaced angles}

\item{method}{method to be used in the estimation of the angular dependence function: "hill" to use the Hill estimator, "cl" for the composite likelihood estimator}

\item{rc_origin}{matrix containing the return curve estimates in the original margins}

\item{q}{quantile to be used for the Hill estimator and/or the Heffernan and Tawn conditional extremes model; default set to 0.95}

\item{k}{polynomial degree for the Bernstein-Bezier polynomials used in the estimation of the angular dependence function using the composite likelihood method; default set to 7}

\item{constrained}{indicates whether or not to incorporate knowledge of the conditional extremes parameters; default set to "no"}

\item{blocksize}{size of the blocks for the block bootstrap; default to 1 for a standard bootstrap approach}

\item{nboot}{number of bootstrap samples; default to 250}

\item{nangles}{number of angles \eqn{m} in the \eqn{(0, \pi/2)} interval; default is set to 150}

\item{alpha}{significance level to compute the confidence intervals}
}
\value{
return a list with estimates for the median and lower and upper bounds of the CI for the empirical probability of lying in a survival region
}
\description{
computes the goodness of fir of the return curve estimates
}
\details{
to do
}
\examples{
library(ReturnCurves)

}
\references{
to do
}
