% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returncurve.R
\docType{methods}
\name{curvetransf}
\alias{curvetransf}
\title{Return Curve in Original margins}
\usage{
curvetransf(curvedata, data, q = 0.95)
}
\arguments{
\item{curvedata}{matrix containing an object of function \code{\link{rc_est}}}

\item{data}{matrix containing the data in the original margins}
}
\value{
return curve estimation in original margins
}
\description{
computes the return curve estimation in the original margins following the equation in the paper
}
\examples{
library(ReturnCurves)

}
\references{
to do
}
