% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_uncertainty.R
\docType{methods}
\name{rc_unc}
\alias{rc_unc}
\title{Uncertainty of the Return Curve estimates}
\usage{
rc_unc(
  data,
  w = seq(0, 1, by = 0.01),
  p,
  method = c("hill", "cl"),
  qmarg = 0.95,
  q = 0.95,
  qalphas = 0.95,
  k = 7,
  constrained = FALSE,
  blocksize = 1,
  nboot = 250,
  nangles = 150,
  alpha = 0.05
)
}
\arguments{
\item{data}{A matrix which contains the data on the original margins.}

\item{w}{Sequence of angles between 0 and 1. Default is \code{seq(0, 1, by = 0.01)}.}

\item{p}{Curve survival probability.}

\item{method}{String that indicates which method is used for the estimation of the angular dependence function. Must either be \code{"hill"}, to use the Hill estimator \insertCite{Hill1975}{ReturnCurves}, or \code{"cl"} to use the composite likelihood estimator approaches. More details can be found in \code{\link{adf_est}}.}

\item{qmarg}{Marginal quantile used to fit the Generalised Pareto Distribution. Default is 0.95.}

\item{q}{\loadmathjax{} Marginal quantile used for the min-projection variable \mjeqn{T^'}{} at angle \mjeqn{\omega}{} \mjeqn{\left(t^'_\omega = t_\omega - u_\omega \mid t_\omega > u_\omega\right)}{}, and/or Hill estimator \insertCite{Hill1975}{ReturnCurves}. Default is 0.95}

\item{qalphas}{Marginal quantile used for the Heffernan and Tawn conditional extremes model \insertCite{HeffernanTawn2004}{ReturnCurves}. Default set to 0.95.}

\item{k}{Polynomial degree for the Bernstein-Bezier polynomials used for the estimation of the angular dependence function using the composite likelihood method. Default set to 7.}

\item{constrained}{Logical. If FALSE (default) no knowledge of the conditional extremes parameters is incorporated in the angular dependence function estimation.}

\item{blocksize}{Size of the blocks for the block bootstrap procedure. If 1 (default), then a standard bootstrap approach is applied.}

\item{nboot}{Number of bootstrap samples to be taken. Default is 250 samples.}

\item{nangles}{\loadmathjax{} Number of angles \mjeqn{m}{m} in the interval \mjeqn{(0, \pi/2)}{} \insertCite{MurphyBarltropetal2023}{ReturnCurves}. Default is 150 angles.}

\item{alpha}{\loadmathjax{} Significance level to compute the \mjeqn{(1-\alpha)}{} confidence intervals. Default is 0.05.}
}
\value{
Returns a list containing: \describe{
\item{median}{A vector containing the median estimates of the return curve.} 
\item{mean}{A vector containing the mean estimates of the return curve.} 
\item{lower}{A vector containing the lower bound of the confidence interval.}
\item{upper}{A vector containing the upper bound of the confidence interval.}
}
}
\description{
Uncertainty assessment of the return curve estimates following the procedure of \insertCite{MurphyBarltropetal2023;textual}{ReturnCurves}.
}
\details{
\loadmathjax{} Define a set of angles \mjdeqn{\boldsymbol{\Theta}:= \left\lbrace \frac{\pi(m+1-j)}{2(m+1)} \mid 1\leq j\leq m\right\rbrace}{} and \mjeqn{L_\theta:=\left\lbrace(x,y)\in R^2_+ \mid \tan(\theta)=y/x\rbrace\right.}{}
For each \mjeqn{\theta\in \boldsymbol{\Theta},}{} \mjeqn{L_\theta}{} intersects the estimated RC\mjeqn{p} exactly once, i.e., \mjeqn{\lbrace\hat{x}_\theta, \hat{y}_\theta\rbrace:= \hat{RC}(p)\cap L_\theta.}{} 
Uncertainty of the return curve is then quantified by the distribution of \mjeqn{\hat{d}_\theta:=\left(\hat{x}^2_\theta + \hat{y}^2_\theta\right)^{1/2}}{} via a (block) bootstrap procedure. More details can be found in \insertCite{MurphyBarltropetal2023}{ReturnCurves}
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(100), runif(100))

dataexp <- margtransf(data)

prob <- 0.001

rc <- rc_est(data = dataexp, p = prob, method = "hill")

rc_orig <- curvetransf(curvedata = rc, data = data)

unc <- rc_unc(data = data, p = prob, method = "hill")


\dontrun{
plot(data, xlab = "X", ylab = "Y", pch = 16)
lines(rc_orig, lwd = 2, col = 2)
lines(unc$median, lwd = 2, col = "orange") # to plot median estimates
lines(unc$mean, lwd = 2, col = "orange") # to plot mean estimates
lines(unc$upper, lty = 'dashed', col = 'blue', lwd = 2)
lines(unc$lower, lty = 'dashed', col = 'blue', lwd = 2)
}

}
\references{
\insertAllCited{}
}
