% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_est.R
\docType{methods}
\name{rc_est}
\alias{rc_est}
\title{Return Curve estimation}
\usage{
rc_est(
  data,
  w = seq(0, 1, by = 0.01),
  p,
  method = c("hill", "cl"),
  q = 0.95,
  qalphas = 0.95,
  k = 7,
  constrained = FALSE
)
}
\arguments{
\item{data}{A matrix containing the data on standard exponential margins.}

\item{w}{Sequence of angles between 0 and 1. Default is \code{seq(0, 1, by = 0.01)}.}

\item{p}{Curve survival probability.}

\item{method}{String that indicates which method is used for the estimation of the angular dependence function. Must either be \code{"hill"}, to use the Hill estimator \insertCite{Hill1975}{ReturnCurves}, or \code{"cl"} to use the composite maximum likelihood estimator. More details can be found in \code{\link{adf_est}}.}

\item{q}{\loadmathjax{} Marginal quantile used for the min-projection variable \mjeqn{T^'}{} at angle \mjeqn{\omega}{} \mjeqn{\left(t^'_\omega = t_\omega - u_\omega \mid t_\omega > u_\omega\right)}{}, and/or Hill estimator \insertCite{Hill1975}{ReturnCurves}. Default is 0.95.}

\item{qalphas}{Marginal quantile used for the Heffernan and Tawn conditional extremes model \insertCite{HeffernanTawn2004}{ReturnCurves}. Default set to 0.95.}

\item{k}{Polynomial degree for the Bernstein-Bezier polynomials used for the estimation of the angular dependence function with the composite likelihood method \insertCite{MurphyBarltropetal2023}{ReturnCurves}. Default set to 7.}

\item{constrained}{Logical. If FALSE (default) no knowledge of the conditional extremes parameters is incorporated in the angular dependence function estimation.}
}
\value{
A matrix containing the estimates of the Return Curve on standard exponential margins.
}
\description{
\loadmathjax{} Estimation of the \mjeqn{p}{p}-probability return curve following \insertCite{MurphyBarltropetal2023;textual}{ReturnCurves}.
}
\details{
\loadmathjax{} Given a probability \mjeqn{p}{p} and a joint survival function \mjeqn{Pr(X>x, Y>y)}{}, 
the \mjeqn{p}{p}-probability return curve is defined as 
\mjdeqn{RC(p):=\left\lbrace(x, y) \in R^2: Pr(X>x, Y>y)=p\right\rbrace.}{} 

\mjeqn{Pr(X>x, Y>y)}{} is estimated using the angular dependence function \mjeqn{\lambda(\omega)}{} introduced by \insertCite{WadsworthTawn2013;textual}{ReturnCurves}. More details on how to estimate \mjeqn{\lambda(\omega)}{} can be found in \code{\link{adf_est}}.
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(100), runif(100))

dataexp <- margtransf(data)

prob <- 0.001

rc <- rc_est(data = dataexp, p = prob, method = "hill")

\dontrun{
plot(dataexp, pch = 20, main = "Return Curve in exponential margins")
lines(rc, col = 2, lwd = 2)
}

}
\references{
\insertAllCited{}
}
