% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_orig.R
\docType{methods}
\name{rc_est}
\alias{rc_est}
\title{Return Curve estimation}
\usage{
rc_est(
  data,
  qmarg = 0.95,
  w = seq(0, 1, by = 0.01),
  p,
  method = c("hill", "cl"),
  q = 0.95,
  qalphas = 0.95,
  k = 7,
  constrained = FALSE,
  tol = 0.001
)
}
\arguments{
\item{data}{A matrix containing the data on the original margins.}

\item{qmarg}{Marginal quantile to be used for the fit of the Generalised Pareto Distribution. Default is \code{0.95}.}

\item{w}{Sequence of angles between \code{0} and \code{1}. Default is \code{seq(0, 1, by = 0.01)}.}

\item{p}{\loadmathjax{} Curve survival probability. Must be \mjeqn{p < 1-q}{p < 1-q} and \mjeqn{p < 1-q_\alpha}{p < 1-qalphas}.}

\item{method}{String that indicates which method is used for the estimation of the angular dependence function. Must either be \code{"hill"}, to use the Hill estimator \insertCite{Hill1975}{ReturnCurves}, or \code{"cl"} to use the composite maximum likelihood estimator.}

\item{q}{\loadmathjax{} Marginal quantile used for the min-projection variable \mjeqn{T^1}{} at angle \mjeqn{\omega}{} \mjeqn{\left(t^1_\omega = t_\omega - u_\omega | t_\omega > u_\omega\right)}{}, and/or Hill estimator \insertCite{Hill1975}{ReturnCurves}. Default is \code{0.95}.}

\item{qalphas}{Marginal quantile used for the Heffernan and Tawn conditional extremes model \insertCite{HeffernanTawn2004}{ReturnCurves}. Default set to \code{0.95}.}

\item{k}{Polynomial degree for the Bernstein-Bezier polynomials used for the estimation of the angular dependence function with the composite likelihood method \insertCite{MurphyBarltropetal2024}{ReturnCurves}. Default set to \code{7}.}

\item{constrained}{Logical. If \code{FALSE} (default) no knowledge of the conditional extremes parameters is incorporated in the angular dependence function estimation.}

\item{tol}{Convergence tolerance for the composite maximum likelihood procedure. Default set to \code{0.0001}.}
}
\value{
A matrix containing the estimates of the Return Curve.
}
\description{
\loadmathjax{} Estimation of the \mjeqn{p}{p}-probability return curve following \insertCite{MurphyBarltropetal2023;textual}{ReturnCurves}.
}
\details{
\loadmathjax{} Let \mjeqn{X, Y\sim Exp(1)}{}. Given a probability \mjeqn{p}{p} and a joint survival function \mjeqn{Pr(X>x, Y>y)}{}, 
the \mjeqn{p}{p}-probability return curve is defined as 
\mjdeqn{RC(p):=\left\lbrace(x, y) \in R^2: Pr(X>x, Y>y)=p\right\rbrace.}{} 

\mjeqn{Pr(X>x, Y>y)}{} is estimated using the angular dependence function \mjeqn{\lambda(\omega)}{} introduced by \insertCite{WadsworthTawn2013;textual}{ReturnCurves}. More details on how to estimate \mjeqn{\lambda(\omega)}{} can be found in \code{\link{adf_est}}.

The return curve estimation \mjeqn{\hat{RC}(p)}{} is done in standard exponential margins and then back transformed onto the original margins.
}
\note{
The parameter \code{qmarg} should be the same as the one used in \code{\link{margtransf}}.
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(1000), rnorm(1000))

n <- dim(data)[1]

prob <- 1/n

rc_orig <- rc_est(data = data, p = prob, method = "hill")

\dontrun{
plot(data, pch = 20, main = "Return Curve on the original margins")
lines(rc_orig, col = 2, lwd = 2)
}

}
\references{
\insertAllCited{}
}
