% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvetransf.R
\docType{methods}
\name{curvetransf}
\alias{curvetransf}
\title{Return Curve estimates in original margins}
\usage{
curvetransf(curvedata, data, q = 0.95)
}
\arguments{
\item{curvedata}{A matrix or data frame containing an object of function \code{\link{rc_est}}}

\item{data}{A matrix or data frame containing the data in the original margins}

\item{q}{Marginal quantile to be used for the fit of the Generalised Pareto Distribution. Default is 0.95.}
}
\value{
A matrix or data frame containing the estimates of the Return Curve in the original margins.
}
\description{
Back transformation of the return curve estimates onto the original margins, following \cite{m2023}
}
\examples{
library(ReturnCurves)

# Generating data for illustration purposes
set.seed(321)
data <- cbind(rnorm(100), runif(100))

dataexp <- margtransf(data)

rc <- rc_est(data = dataexp, p = 1e^-3, method = "hill")

\dontrun{
plot(dataexp, pch = 20, main = "Return Curve in exponential margins")
lines(rc, col = 2, lwd = 2)
}

rc_orig <- curvetransf(curvedata = rc, data = data)

\dontrun{
plot(data, pch = 20, main = "Return Curve in original margins")
lines(rc_origin, col = 2, lwd = 2)
}

}
\references{
\insertRef{MurphyBarltropetal2023}{m2023}
}
